Defined a template, docker config, and started db setup
There's a lot going on in this PR, and sadly, I don't yet have a firm
enough grasp on the goings-on to write any tests. That is certainly a
failing. Once I understand the ways that I can make calls to the db,
I will wrap those as appropriate and write some tests.

Additionally, I will need some tests at the integration level to
guarantee that the generated page works correctly; i.e. clicking a
button increments the counter, etc.