Regression fix for DB test was not deterministic
Now the test succeeds every time, but I had to set a flag on the test
to not watch the resource table for leaks, which is what was causing the
intermittent failures. I am not sure when, but it seems that a tcpstream
is sometimes opened and not closed during the course of the test.