the tests run, the app has no functionality
I got myself in such a pickle trying to follow other peoples' code
that I ended up painting myself into a corner with colors and brushes
that I did not know how to use.
I know how to use TDD, so from here on out, that is my sword and shield
in the fight to make sense of this app.