Added a seed file, and a set up to test with the database
I wrote some deno process execution code to rollback, migrate, and seed
the local database whenever the 'server.test.ts' is run.

Also, my intellisense (coc.nvim with ts and deno plugins) complains
that I don't need the 'await' keyword on the requests to the superoak
testing framework, but failing to use await results in some very odd
behavior, including (seems to me) NON-DETERMINISTIC tests. Eegad.
So be sure to leave that 'await request.method' in there, even when it
complains that it isn't needed.