separated DB connections for the message store and 'normal' RDBMS
Made changes to the config and utility functions (to draw from the env
variables) so that I can fully separated the concerns of the message
store and the 'regular' RDBMS postgres instance.

I have also stubbed out the record-viewings app, which will be a simple
first try for the message_store.

To be sure that I could test that the message_store comes up when the
app launches, my config is currently exposing a raw connection to the
message_store postgres instance. Sooo...

TODO: Once the application can read/ consume messages from the
message_store, the raw database variable exposed through the config for
the message_store postgres instance should be removed.