tests and refactors for the message writing process
I elected to make the three value 'write' method for the messageStore
take an object with named props so that the function is
self-documenting.
I think that this would be a good choice going forward for most
functions that take three or more values as arguments.

The tests for the messageStore are still using a raw reference to the
postgres client connection to perform their tests. Once there is a
method for reading from the messageStore, that raw connection needs to
be taken away to limit the surface area of the messageStore/ config API.