breaking changes, that are currently broken :)
I am in the process of switching the entire data model over to the
messageStore (messageDB) being the single source of truth, and having
never dealt with this architecture before, I am a bit overwhelmed.
As such, I couldn't wrap my head around how to write tests for this
except my highest-level test to check the view-data on the '/' page.

Once I understand the system better (which I think will be soon), I will
try again to write tests for the more atomic pieces of the message store
interaction.